+++ src/pages/Cphilipse/Cphilipse.tsx
@@ -1,30 +1,15 @@
-import React, {useCallback, useEffect} from 'react';
-import {Linking, ScrollView, Text, View} from 'react-native';
+import React, {useEffect} from 'react';
+import {ScrollView, View} from 'react-native';
 import styles from './cphilipse.style';
 import BackButton from '../../components/BackButton';
-import Basketball from '../../components/Basketball';
-import Animated, {
-  useAnimatedGestureHandler,
-  useAnimatedStyle,
-  useSharedValue,
-  withDecay,
-  withRepeat,
-  withSpring,
-  withTiming,
-} from 'react-native-reanimated';
-import {colors, images, metrics} from '../../themes';
+import {images} from '../../themes';
 import Header from '../../components/Header';
 import {useBounce} from './animations/useBounce';
 import {useCrucifix} from './animations/useCrucifix';
-import Crucifix from '../../components/Crucifix/Crucifix';
-import Button from '../../components/Button';
-import {PanGestureHandler} from 'react-native-gesture-handler';
 import {swipeBasketball} from './animations/swipeBasketball';
-import Icons from '../../enum/Icons';
-import Icon from 'react-native-vector-icons/Entypo';
-import {ICON_SIZE} from '../../components/Icon/Icon';
 import Paragraph from '../../components/Paragraph';
 import {paragraphs} from '../../utils/DummyData';
+import Items from './components/Items';
 
 interface Props {
   navigation: any;
@@ -34,7 +19,6 @@ interface Props {
  * TODO: Make buttons padding bigger, so it'll be easier to press.
  * TODO: Locales for the text.
  * TODO: Abstract the body.
- * TODO: Abstract the character container.
  * TODO: Test this page on multiple device simulators.
  * TODO: Create 'master' branch, change 'main' to 'master' branch. Delete 'main' branch.
  *  Merge all to 'dev' and the 'master' branch. Remove 'setup' branch.
@@ -57,55 +41,23 @@ const Cphilipse = ({navigation}: Props) => {
     startAnimatingCrucifix();
   }, [startBasketballAnimation, startAnimatingCrucifix]);
 
-  const openLinkedIn = useCallback(() => {
-    Linking.openURL('https://www.linkedin.com/in/clemens-philipse-2615b9162/');
-  }, []);
-  const openGitHub = useCallback(() => {
-    Linking.openURL('https://github.com/CPhilipse/CPhilipse_APP');
-  }, []);
-
   return (
     <View style={styles.container}>
       <ScrollView>
         <Header title={'Clemens Philipse'} image={images.cphilipse} />
-
-        <PanGestureHandler {...{onGestureEvent}}>
-          <Animated.View style={[style, styles.basketball, gestureStyle]}>
-            <Basketball />
-          </Animated.View>
-        </PanGestureHandler>
-        <View style={styles.crucifix}>
-          <Crucifix
-            {...{
-              scaleStyle,
-              colorStyle,
-              colorStyle2,
-              colorStyle3,
-              colorStyle4,
-              colorStyle5,
-            }}
-          />
-        </View>
-
-        <View style={styles.characterColumn}>
-          <View style={[styles.gameController]}>
-            <Icon
-              name={Icons.GAME_CONTROLLER}
-              size={ICON_SIZE}
-              color={colors.red}
-            />
-          </View>
-          <View style={[styles.nineteenContainer]}>
-            <Text style={styles.nineteen}>19</Text>
-          </View>
-          <Button onPress={openLinkedIn} style={styles.button}>
-            <Text style={styles.buttonText}>LinkedIn</Text>
-          </Button>
-          <Button onPress={openGitHub} style={styles.button}>
-            <Text style={styles.buttonText}>GitHub</Text>
-          </Button>
-        </View>
-
+        <Items
+          {...{
+            style,
+            gestureStyle,
+            onGestureEvent,
+            colorStyle,
+            colorStyle2,
+            colorStyle3,
+            colorStyle4,
+            colorStyle5,
+            scaleStyle,
+          }}
+        />
         <Paragraph paragraph={paragraphs} />
         <BackButton onPress={() => navigation.navigate('Home')} />
       </ScrollView>
