+++ src/pages/Cphilipse/components/Items/index.ts
@@ -1,0 +1,3 @@
+import Items from './Items';
+
+export default Items;
